requirements.txt
langchain~=0.3.14
gigachain-community~=0.2.16.post3
sentence-transformers~=3.3.1
httpx~=0.28.1
aiofiles~=24.1.0
python-dotenv~=1.0.1
chromadb~=0.6.1
fastapi~=0.115.6
pydantic~=2.10.4
pydantic-settings~=2.7.1
elasticsearch~=8.17.0